NRL Service (#6)
* feat: added fetching fixtures from nrl

* test: unit tests to fetch fixtures from past games

* fix: change from int32 to int64 to handle the larger nrl ids

* fix: need to allow setting the fixture id to the nrl match id

* feat: data integration for the nrl api

* test: test the storing nrl data in the db

* feat: working scheduled fetch and store

* fix: teams have a competition id which make it easier to tell them apart

+++ backend/internal/db/competitions.sql.go
@@ -14,7 +14,7 @@ SELECT id, name FROM competitions WHERE id = $1
 `
 
 // Retrieve a specific competition by its unique identifier.
-func (q *Queries) GetCompetitionByID(ctx context.Context, id int32) (*Competition, error) {
+func (q *Queries) GetCompetitionByID(ctx context.Context, id int64) (*Competition, error) {
 	row := q.db.QueryRow(ctx, getCompetitionByID, id)
 	var i Competition
 	err := row.Scan(&i.ID, &i.Name)
