+++ backend/internal/db/migration/000001_create_competitions_table.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE competitions (
-  id SERIAL PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   name VARCHAR(255) NOT NULL
 );
